Removed the clients and sessions to requestHandler class

+++ src/SIP/Session.cpp
@@ -1,6 +1,6 @@
 #include "Session.hpp"
 
-Session::Session(std::string callID, std::shared_ptr<SipClient> src, size_t srcRtpPort) :
+Session::Session(std::string callID, std::shared_ptr<SipClient> src, uint32_t srcRtpPort) :
 	_callID(std::move(callID)), _src(src), _state(State::Invited), _srcRtpPort(srcRtpPort), _destRtpPort(0)
 {
 }
@@ -17,7 +17,7 @@ void Session::setState(State state)
 	}
 }
 
-void Session::setDest(std::shared_ptr<SipClient> dest, size_t destRtpPort)
+void Session::setDest(std::shared_ptr<SipClient> dest, uint32_t destRtpPort)
 {
 	_dest = dest;
 	_destRtpPort = destRtpPort;
