+++ src/SIP/Session.hpp
@@ -17,10 +17,10 @@ public:
 	};
 
 
-	Session(std::string callID, std::shared_ptr<SipClient> src, size_t srcRtpPort);
+	Session(std::string callID, std::shared_ptr<SipClient> src, uint32_t srcRtpPort);
 	
 	void setState(State state);
-	void setDest(std::shared_ptr<SipClient> dest, size_t destRtpPort);
+	void setDest(std::shared_ptr<SipClient> dest, uint32_t destRtpPort);
 
 	std::string getCallID() const;
 	std::shared_ptr<SipClient> getSrc() const;
@@ -32,8 +32,8 @@ private:
 	std::shared_ptr<SipClient> _dest;
 	State _state;
 
-	int _srcRtpPort;
-	int _destRtpPort;	
+	uint32_t _srcRtpPort;
+	uint32_t _destRtpPort;
 };
 
 #endif
